# 2/first-frontend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "0.1"
    },
    "paths": {
...
        "/inchirieri": {
            "get": {
                "tags": [
                    "Inchirieri"
                ],
                "summary": "Returneaza toate inchirierile",
                "description": "Returneaza toate inchirierile",
                "operationId": "29ae57419888b98b0caba889bd667970",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "put": {
                "tags": [
                    "Inchirieri"
                ],
                "summary": "Update inchirierea cu id-ul dat",
                "description": "Update inchirierea cu id-ul dat",
                "operationId": "55db621f52d27ccbde20a8825ace63c8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "idInchiriere": {
                                        "description": "",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "idClient": {
                                        "description": "",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "idMasina": {
                                        "description": "",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "dataInchiriere": {
                                        "description": "",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2022-10-17"
                                    },
                                    "dataPredareLimita": {
                                        "description": "",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2022-10-23"
                                    },
                                    "dataPredareEfectiva": {
                                        "description": "",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2022-10-23"
                                    },
                                    "idLocatieInchiriere": {
                                        "description": "",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "idLocatiePredare": {
                                        "description": "",
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Inchirieri"
                ],
                "summary": "Adauga inchirierea",
                "description": "Adauga inchirierea",
                "operationId": "f2a3b0c34553294d5f6a12ebe1129eed",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "idClient": {
                                        "description": "",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "idMasina": {
                                        "description": "",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "dataInchiriere": {
                                        "description": "",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2022-10-17"
                                    },
                                    "dataPredareLimita": {
                                        "description": "",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2022-10-23"
                                    },
                                    "dataPredareEfectiva": {
                                        "description": "",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2022-10-23"
                                    },
                                    "idLocatieInchiriere": {
                                        "description": "",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "idLocatiePredare": {
                                        "description": "",
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/inchirieri/{id}": {
            "get": {
                "tags": [
                    "Inchirieri"
                ],
                "summary": "Returneaza inchirierea cu id-ul dat",
                "description": "Returneaza inchirierea cu id-ul dat",
                "operationId": "896f541c60cf3169df364a3289be6ad5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id-ul inchirierii dorit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 30
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Inchirieri"
                ],
                "summary": "Sterge inchirierea cu id-ul dat",
                "description": "Sterge inchirierea cu id-ul dat",
                "operationId": "248c1f7752815d2d3a3c155f5c0f2f53",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id-ul inchirierii dorite",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 30
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    }
}